// src/frontend/react-admin/src/generated/models/visite-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserResponse } from './user-response';
import { VisiteurResponse } from './visiteur-response';
/**
 * 
 * @export
 * @interface VisiteResponse
 */
export interface VisiteResponse {
    /**
     * 
     * @type {number}
     * @memberof VisiteResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof VisiteResponse
     */
    motif?: string;
    /**
     * 
     * @type {string}
     * @memberof VisiteResponse
     */
    heureDebut?: string;
    /**
     * 
     * @type {string}
     * @memberof VisiteResponse
     */
    heureFin?: string;
    /**
     * 
     * @type {Date}
     * @memberof VisiteResponse
     */
    dateVisite?: Date;
    /**
     * 
     * @type {string}
     * @memberof VisiteResponse
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof VisiteResponse
     */
    status?: string;
    /**
     * 
     * @type {UserResponse}
     * @memberof VisiteResponse
     */
    user?: UserResponse;
    /**
     * 
     * @type {VisiteurResponse}
     * @memberof VisiteResponse
     */
    visiteur?: VisiteurResponse;
    /**
     * 
     * @type {Date}
     * @memberof VisiteResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof VisiteResponse
     */
    updatedAt?: Date;
}